Changed default decimal type from float to double for higher precision

+++ include/cartesian.h
@@ -4,8 +4,8 @@
 #include "application_params.h"
 struct Point
 {
-    float x = 0, y = 0, z = 0;
-    Point(float x = 0, float y = 0, float z = 0) : x(x), y(y), z(z) {}
+    double x = 0, y = 0, z = 0;
+    Point(double x = 0, double y = 0, double z = 0) : x(x), y(y), z(z) {}
 
     Point operator+(const Point &point) const
     {
@@ -31,7 +31,7 @@ struct Point
     }
 
     Point
-    operator/(const float &num) const
+    operator/(const double &num) const
     {
         return {
             x / num,
@@ -39,7 +39,7 @@ struct Point
             z / num};
     }
 
-    Point operator*(const float &num) const
+    Point operator*(const double &num) const
     {
         return {
             x * num,
@@ -49,7 +49,7 @@ struct Point
 
     void normalize()
     {
-        float len = sqrt(x * x + y * y + z * z);
+        double len = sqrt(x * x + y * y + z * z);
         if (len == 0)
             return;
         x /= len;
@@ -58,7 +58,7 @@ struct Point
             z /= len;
     }
 
-    float magnitude()
+    double magnitude()
     {
         return sqrt(x * x + y * y + z * z);
     }
@@ -68,8 +68,8 @@ struct Line
 {
     // x exists as a means to tell if the line is completely vertical or not
     bool isVertical = false;
-    float m = 0, x = 0, c = 0;
-    const float VERTICAL_THRESHOLD = 0.00001;
+    double m = 0, x = 0, c = 0;
+    const double VERTICAL_THRESHOLD = 0.00001;
 
     //  Constructs a line that passes through two points.
     Line(const Point &p1, const Point &p2)
@@ -79,8 +79,8 @@ struct Line
         // p1.y = m*p1.x + c
         // p1.y - (m*p1.x) = c
 
-        float rise = p1.y - p2.y;
-        float run = p1.x - p2.x;
+        double rise = p1.y - p2.y;
+        double run = p1.x - p2.x;
 
         // Only have to check if run is too small to prevent dividing by zero errors (m = rise/run)
         if (abs(run - 0) < VERTICAL_THRESHOLD)
@@ -103,14 +103,14 @@ struct Line
     }
 
     // Constructs a vertical line
-    Line(float x)
+    Line(double x)
     {
         this->x = x;
         isVertical = true;
     }
 
     // Constructs an ordinary line (non-vertical)
-    Line(float m, float c)
+    Line(double m, double c)
     {
         this->m = m;
         this->c = c;
@@ -118,7 +118,7 @@ struct Line
     }
 
     // Constructs a line with gradient 'm', that intersects with a point 'p'
-    Line(float m, Point p)
+    Line(double m, Point p)
     {
         // y = mx + c
         // p.y = m*p.x + c
@@ -172,7 +172,7 @@ namespace Math
     }
 
     // A cross product of two, 2D vectors is also known as a 'wedge product'
-    static float crossProd2D(const Point &p1, const Point &p2)
+    static double crossProd2D(const Point &p1, const Point &p2)
     {
         return p1.x * p2.y - p1.y * p2.x;
     }
@@ -186,25 +186,25 @@ namespace Math
             p2.x * p1.y - p2.y * p1.x};
     }
 
-    static float dotProd(const Point &p1, const Point &p2);
-    static float dotProd(const Point &p1, const Point &p2)
+    static double dotProd(const Point &p1, const Point &p2);
+    static double dotProd(const Point &p1, const Point &p2)
     {
         return (p1.x * p2.x) + (p1.y * p2.y) + (p1.z * p2.z);
     }
 
     // See https://proofwiki.org/wiki/Square_of_Vector_Cross_Product
-    static float crossProdSquare(Point p1, Point p2)
+    static double crossProdSquare(Point p1, Point p2)
     {
         // square of a vector cross product axb equals:
         // (a.a)(b.b) - (a.b)^2 where '.' is the dot product
-        float a_2 = dotProd(p1, p1);
-        float b_2 = dotProd(p2, p2);
-        float ab = dotProd(p1, p2);
+        double a_2 = dotProd(p1, p1);
+        double b_2 = dotProd(p2, p2);
+        double ab = dotProd(p1, p2);
         return a_2 * b_2 - pow(ab, 2);
     }
 
-    static float dist(const Point &p1, const Point &p2);
-    static float dist(const Point &p1, const Point &p2)
+    static double dist(const Point &p1, const Point &p2);
+    static double dist(const Point &p1, const Point &p2)
     {
         return sqrt(
             pow((p1.x - p2.x), 2) +
@@ -213,9 +213,10 @@ namespace Math
 
     // Calculates the instantaneous velocity of a point "p" rotating about an axis "c"
     // Does not take into account any translational motion of the current point
-    static Point instantVelRot2D(const Point &p, const Point &c, float rot)
+    static Point instantVelRot2D(const Point &p, const Point &c, double rot)
     {
         Point r_cp = p - c;
+        // The rotational axis in 2 dimensions is always perpendicular to the x-y plane 
         return crossProd3D(r_cp, {0, 0, rot});
     }
 
@@ -243,7 +244,7 @@ namespace Math
             return {l2.x, l1.m * l2.x + l1.c};
         }
 
-        float x_intersection = (l2.c - l1.c) / (l1.m - l2.m);
+        double x_intersection = (l2.c - l1.c) / (l1.m - l2.m);
         return {x_intersection, l1.m * x_intersection + l1.c};
     }
 }
